+++ src/js/main.js
@@ -115,6 +115,14 @@
             } else{
                 hidePopups();
             }
+
+            if ($.inArray('actionbar-selectall-link', event.target.classList) > -1) {
+                event.preventDefault();
+                $('.link-outer').each(function () {
+                    console.log('lol');
+                    toggleLinkSelection($(this), true);
+                });
+            }
         });
 
         $('.popup-trigger').on('click', function(){
@@ -730,6 +738,7 @@
      *                             - noHtmlEscape {Boolean} Prevent HTML escape of title and body.
      *                             - value {String} Value of input field when using type prompt.
      *                             - buttonLabelOk {String} Label for OK button. Defaults to 'OK'.
+     *                             - buttonClassesOk {String} Classes for OK button.
      * @return {Void}
      */
     var displayModal = function(title, text, type, callback, options){
@@ -749,11 +758,11 @@
                 footer += '<button class="button ripple pull-right modal-ok">' + (options && options.buttonLabelOk ? options.buttonLabelOk : 'OK') + '</button>';
                 break;
             case 'confirm':
-                footer += '<button class="button ripple button-alert pull-right modal-ok">' + (options && options.buttonLabelOk ? options.buttonLabelOk : 'OK') + '</button><button class="button ripple pull-right modal-cancel">Cancel</button>';
+                footer += '<button class="button ripple ' + (options && options.buttonClassesOk ? options.buttonClassesOk : 'button-alert') + ' pull-right modal-ok">' + (options && options.buttonLabelOk ? options.buttonLabelOk : 'OK') + '</button><button class="button ripple pull-right modal-cancel">Cancel</button>';
                 break;
             case 'prompt':
                 body += '<input type="text" class="input-new-tag" placeholder="Enter a new value..." value="' + options.value + '"/>';
-                footer += '<button class="button ripple button-primary pull-right modal-ok">' + (options && options.buttonLabelOk ? options.buttonLabelOk : 'OK') + '</button><button class="button ripple pull-right modal-cancel">Cancel</button>';
+                footer += '<button class="button ripple ' + (options && options.buttonClassesOk ? options.buttonClassesOk : 'button-primary') + ' pull-right modal-ok">' + (options && options.buttonLabelOk ? options.buttonLabelOk : 'OK') + '</button><button class="button ripple pull-right modal-cancel">Cancel</button>';
                 break;
             default:
                 console.log('Modal type must be alert, confirm or prompt. ' + type + ' isn\'t recognized.');
@@ -797,7 +806,7 @@
         }
 
         var uid = guid();
-        var html = '<div id="' + uid + '" class="hidden actionbar ' + options.classes + '"><div class="container"><div class="row"><div class="actionbar-label">' + options.label + '</div><div class="actionbar-controls">';
+        var html = '<div id="' + uid + '" class="hidden actionbar ' + options.classes + '"><div class="container"><div class="row"><div class="actionbar-label">' + options.label + '</div><div class="actionbar-selectall">- <a href="#" class="actionbar-selectall-link">select all</a></div><div class="actionbar-controls">';
 
         if(options.displayCancel){
             html += '<button type="button" class="button button-default" id="actionbar-cancel">Cancel</button>';
@@ -841,6 +850,54 @@
         localStorage.setItem('expand', isExpanded);
     };
 
+    var toggleLinkSelection = function ($link, forceSelect, forceDeselect) {
+        var id = $link.attr('id');
+
+        var select = function () {
+            $link.addClass('is-selected');
+            model.selectedLinks[id] = {
+                id: id,
+                title: escapeHtml($link.find('.link-title').text())
+            };
+
+            refreshActionBarLabel();
+        }
+
+        var deselect = function () {
+            $link.removeClass('is-selected');
+            delete model.selectedLinks[id];
+
+            refreshActionBarLabel();
+        }
+
+        if (forceSelect) {
+            select();
+            return;
+        }
+
+        if (forceDeselect) {
+            deselect();
+            return;
+        }
+
+        if($link.hasClass('is-selected')){
+            deselect();
+        } else {
+            select();
+        }
+    };
+
+    var refreshActionBarLabel = function() {
+        var numberOfLinksSelected = objectSize(model.selectedLinks);
+        $batchDeleteActionBar.find('.actionbar-label').text(numberOfLinksSelected + ' links selected');
+
+        if (numberOfLinksSelected === 0) {
+            $batchDeleteActionBar.find('button:not(#actionbar-cancel)').attr('disabled', 'disabled');
+        } else {
+            $batchDeleteActionBar.find('button:not(#actionbar-cancel)').removeAttr('disabled');
+        }
+    };
+
     var toggleBatchMode = function(){
         if(!isBatchModeEnabled){
             isBatchModeEnabled = true;
@@ -856,10 +913,6 @@
                 }, {
                     buttonLabelOk: 'Understood!'
                 });
-            }
-
-            var refreshActionBarLabel = function(){
-                $batchDeleteActionBar.find('.actionbar-label').text(objectSize(model.selectedLinks) + ' links selected');
             }
 
             $('.links-list').addClass('is-selectable');
@@ -923,6 +976,66 @@
                             }, { 
                                 noHtmlEscape: true,
                                 buttonLabelOk: 'Delete ' + length + ' links'
+                            });
+                        }
+                    },
+                    {
+                        id: 'set-links-public-button',
+                        label: 'Set public',
+                        classes: 'button button-primary',
+                        callback: function (event) {
+                            var linksIds = '';
+                            var linksTexts = '<ul class="is-bordered">';
+                            var linksIdTab = [];
+                            var length = objectSize(model.selectedLinks);
+
+                            for(var id in model.selectedLinks){
+                                linksIdTab.push(id);
+                                linksTexts += '<li>#<strong>' + id + '</strong>' + model.selectedLinks[id].title + '</li>';
+                            }
+
+                            linksTexts += '</ul>';
+                            linksIds = linksIdTab.join('+');
+                            var url = '?change_visibility&newVisibility=public&ids='+ linksIds +'&token='+ encodeURIComponent($('#token').val());
+
+                            displayModal('Are you sure to set those ' + length + ' links public?', 'The following links will be set as <strong>public</strong>: ' + linksTexts, 'confirm', function(accepted){
+                                if(accepted){
+                                    window.location.href = url;
+                                }
+                            }, { 
+                                noHtmlEscape: true,
+                                buttonLabelOk: 'Set ' + length + ' links public',
+                                buttonClassesOk: 'button-primary'
+                            });
+                        }
+                    },
+                    {
+                        id: 'set-links-private-button',
+                        label: 'Set private',
+                        classes: 'button button-primary',
+                        callback: function (event) {
+                            var linksIds = '';
+                            var linksTexts = '<ul class="is-bordered">';
+                            var linksIdTab = [];
+                            var length = objectSize(model.selectedLinks);
+
+                            for(var id in model.selectedLinks){
+                                linksIdTab.push(id);
+                                linksTexts += '<li>#<strong>' + id + '</strong>' + model.selectedLinks[id].title + '</li>';
+                            }
+
+                            linksTexts += '</ul>';
+                            linksIds = linksIdTab.join('+');
+                            var url = '?change_visibility&newVisibility=private&ids='+ linksIds +'&token='+ encodeURIComponent($('#token').val());
+
+                            displayModal('Are you sure to set those ' + length + ' links private?', 'The following links will be set as <strong>private</strong>: ' + linksTexts, 'confirm', function(accepted){
+                                if(accepted){
+                                    window.location.href = url;
+                                }
+                            }, { 
+                                noHtmlEscape: true,
+                                buttonLabelOk: 'Set ' + length + ' links private',
+                                buttonClassesOk: 'button-primary'
                             });
                         }
                     }